CS-908 Save a descriptor to Redis first, only then to Mongo
This is needed because Redis may be down more frequently. If we store to Mongo first and Redis is down, the descriptor save operation will fail, and the entity itself will not be saved (because we usually save descriptor before saving its entity), so the descriptor in Mongo will be useless, just a garbage.

- save to Redis first for non-reactive single-save case