CS-667 Port collection top limiting mechanism to reactive context
Our responses may contain embedded collections (instances of CollectionReference) each of them contains some top elements. Such collection tops may be of different size (in elements), each element may render to JSON of a very different size. Given there are heavy collections, we may produce very big responses.

A partial solution to this problem is to estimate the size of the JSON produced by a collection top and, if it exceeds a limit, cut the collection top to a smaller number of elements.

This was already implemented earlier for non-reactive case. Now, this approach is ported to the reactive context.

There are plans to develop this mechanism further.