CS-668 Resolve Descriptor.externalId reactively before serializing to JSON in reactive context
There is a problem in a reactive context. We return a ResponseDto which may contain Descriptor instances. Some of them may be obtained from the database, and so they may have externalId not filled. During serialization to JSON, getExternalId() gets called on Descriptors. If a Descriptor's externalId is not filled, each getExternalId() call will cause a trip to Redis (and, potentially, to Mongo) which will be done in a non-reactive, blocking, way. This ruins reactivity.

So, they idea is to resolve the externalIds of all the Descriptors residing in a ResponseDto manually. This can be done using a Aspect.

Also, we apply an optimization here: we first collect all the Descriptors that need externalId resolved, and then we go to Redis/Mongo in a batch, so effectively we use one fat operation instead of a few (or many) small operations, and we get economy on round-trips.