CS-1152 After collection top is filled, its collections are not processed by makeup when starting with a CollectionReference
The problem we faced was that in Enabler, when listing domains via REST API, collections inside Instance (namely versions and origins) were not made up (id/url/top were not filled).
This was caused by the fact that 'list domains' uses manual collection makeup application (starting with a CollectionReference), and in this mode the makeup was not applied recursively to new objects which appeared after top was filled.
Now, a refactoring was made so that in both modes (automatic and manual) such a recursive makeup application is made.