CS-1213 Read under transaction when saving a versioned entity to Mongo
When adding a new snapshot, we first read current snapshot. This read should be done under the same transaction under which a save will be done.
This required to move from inTransaction() method (which is anyway deprecated), so for now we are trying to use declarative reactive transactions with @Transactional