CS-1270 Store strings in natural form and not as JSON in Redis
Before, JSON codec was used to store even plain strings in Redis, so in our maps, all keys (which are Strings) were surrounded with double quotes. It is more natural to store them as just strings.
Also, this commit removes a default codec (JSON) from Redisson config. It is suggested that codecs are supplied explicitly every time a Redis structure is used.