CS-1291 Change date-time format to ISO-8601 compliant in Extremum
Now, we produce date-time strings like 2020-01-02T23:59:59.123456+0300. The problem is that, according to paragraph 4.3.3d of the ISO-8601 paper, such values are not valid ISO-8601 date-times: a colon in the zone offset is missing.

We must both allow and produce only date-times that are ISO-8601 compliant. Examples are:

2020-01-12T23:59:59.123456+03:00

2020-01-12T23:59:59.123000Z

So the following must be true:

Zone offset must have format like '+03:00', with hours, colon, and minutes, or simply be a literal 'Z' meaning the same as '+00:00'