CS-1302 Return back to saving descriptor to Mongo first, only then to Redis
Initially, we were saving a Descriptor to Mongo first, only then to Redis.

Then, in CS-809 we flipped the order to avoid creation of garbage in MongoDB when Redis starts failing.

Now, the ordering introduced in CS-809 turned out problematic. MongoDB has unique indices, and we rely on them. But if a descriptor is already saved to Redis, it stays there. Moreover, link to it stays in internalDescriptorIds map, so the Descriptor becomes inconsistent: judging by MongoDB, only its first version exists, but in Redis both exist, and the second is referenced, so it is picked.

If we return to the original order of saving (first to MongoDB, then to Redis), this problem goes away.