CS-1381 Use 'uuuu' instead of 'yyyy' in ISO-8601 date format in Extremum
'yyyy' stands for 'era year', so it is, strictly speaking, not enough to reconstruct the year without era designation. This means that in strict contexts parsing with a format like yyyy-MM-dd'T'HH:mm:ss.SSSXXX might fail. It is more robust to use 'uuuu' which stands for proleptic year that does not require era.