CS-1390 Remove equals and hashCode methods from Descriptor
There are cases when we cannot say that descriptors are equal or not without resolving them, which requires to go to a database. And it is an absolute evil to go to a database from a hashCode() method. Plus this does not play well with reactive environments at all.

Such a ‘non-comparable’ descriptors example: one of the descriptors only has externalId, and another one only has internalId loaded.

The clients are better off to compare descriptors themselves if they need this (probably in a wrapper) making sure that the descriptors are fully resolved.